Move Schema only notation to Absinthe.Schema file
To prevent confusion, the notation that can only be used in an
Absinthe.Schema module has been moved to the Absinthe.Schema module.

See e.g.
https://elixirforum.com/t/can-you-create-circular-dependancies-when-using-import-types/51620
