+++ lib/absinthe/schema.ex
@@ -1,5 +1,6 @@
 defmodule Absinthe.Schema do
   alias Absinthe.Type
+  alias Absinthe.Utils
   alias __MODULE__
 
   @type t :: module
@@ -113,6 +114,18 @@ defmodule Absinthe.Schema do
       )
     end
 
+    opts =
+      Keyword.merge(opts,
+        except: [
+          import_types: 1,
+          import_types: 2,
+          import_type_extensions: 1,
+          import_type_extensions: 2,
+          import_directives: 1,
+          import_directives: 2
+        ]
+      )
+
     quote do
       use Absinthe.Schema.Notation, unquote(opts)
       import unquote(__MODULE__), only: :macros
@@ -168,6 +181,7 @@ defmodule Absinthe.Schema do
   @doc """
   Defines a root Query object
   """
+  @doc group: :notation
   defmacro query(raw_attrs \\ [name: @default_query_name], do: block) do
     record_query(__CALLER__, raw_attrs, block)
   end
@@ -198,6 +212,7 @@ defmodule Absinthe.Schema do
   end
   ```
   """
+  @doc group: :notation
   defmacro mutation(raw_attrs \\ [name: @default_mutation_name], do: block) do
     record_mutation(__CALLER__, raw_attrs, block)
   end
@@ -302,6 +317,7 @@ defmodule Absinthe.Schema do
   need, in the event that different groups of mutations return different results
   that require different topic functions.
   """
+  @doc group: :notation
   defmacro subscription(raw_attrs \\ [name: @default_subscription_name], do: block) do
     record_subscription(__CALLER__, raw_attrs, block)
   end
@@ -314,6 +330,262 @@ defmodule Absinthe.Schema do
     env
     |> Absinthe.Schema.Notation.recordable!(:subscription, @placement[:subscription])
     |> Absinthe.Schema.Notation.record!(@object_type, :subscription, attrs, block)
+  end
+
+  @placement {:import_types, [toplevel: true]}
+  @doc """
+  Import types from another module
+
+  Very frequently your schema module will simply have the `query` and `mutation`
+  blocks, and you'll want to break out your other types into other modules. This
+  macro imports those types for use the current module.
+
+  To selectively import types you can use the `:only` and `:except` opts.
+
+  ## Placement
+
+  #{Utils.placement_docs(@placement)}
+
+  ## Examples
+  ```
+  import_types MyApp.Schema.Types
+
+  import_types MyApp.Schema.Types.{TypesA, TypesB}
+
+  import_types MyApp.Schema.Types, only: [:foo]
+
+  import_types MyApp.Schema.Types, except: [:bar]
+  ```
+  """
+  @doc group: :notation
+  defmacro import_types(type_module_ast, opts \\ []) do
+    env = __CALLER__
+
+    type_module_ast
+    |> Macro.expand(env)
+    |> do_import_types(env, opts)
+  end
+
+  defp do_import_types({{:., _, [{:__MODULE__, _, _}, :{}]}, _, modules_ast_list}, env, opts) do
+    for {_, _, leaf} <- modules_ast_list do
+      type_module = Module.concat([env.module | leaf])
+
+      do_import_types(type_module, env, opts)
+    end
+  end
+
+  defp do_import_types(
+         {{:., _, [{:__aliases__, _, [{:__MODULE__, _, _} | tail]}, :{}]}, _, modules_ast_list},
+         env,
+         opts
+       ) do
+    root_module = Module.concat([env.module | tail])
+
+    for {_, _, leaf} <- modules_ast_list do
+      type_module = Module.concat([root_module | leaf])
+
+      do_import_types(type_module, env, opts)
+    end
+  end
+
+  defp do_import_types({{:., _, [{:__aliases__, _, root}, :{}]}, _, modules_ast_list}, env, opts) do
+    root_module = Module.concat(root)
+    root_module_with_alias = Keyword.get(env.aliases, root_module, root_module)
+
+    for {_, _, leaf} <- modules_ast_list do
+      type_module = Module.concat([root_module_with_alias | leaf])
+
+      do_import_types(type_module, env, opts)
+    end
+  end
+
+  defp do_import_types(module, env, opts) do
+    Module.put_attribute(env.module, :__absinthe_type_imports__, [
+      {module, opts} | Module.get_attribute(env.module, :__absinthe_type_imports__) || []
+    ])
+
+    []
+  end
+
+  @placement {:import_directives, [toplevel: true]}
+  @doc """
+  Import directives from another module
+
+  To selectively import directives you can use the `:only` and `:except` opts.
+
+  ## Placement
+  #{Utils.placement_docs(@placement)}
+
+  ## Examples
+  ```
+  import_directives MyApp.Schema.Directives
+
+  import_directives MyApp.Schema.Directives.{DirectivesA, DirectivesB}
+
+  import_directives MyApp.Schema.Directives, only: [:foo]
+
+  import_directives MyApp.Schema.Directives, except: [:bar]
+  ```
+  """
+  @doc group: :notation
+  defmacro import_directives(type_module_ast, opts \\ []) do
+    env = __CALLER__
+
+    type_module_ast
+    |> Macro.expand(env)
+    |> do_import_directives(env, opts)
+  end
+
+  defp do_import_directives({{:., _, [{:__MODULE__, _, _}, :{}]}, _, modules_ast_list}, env, opts) do
+    for {_, _, leaf} <- modules_ast_list do
+      type_module = Module.concat([env.module | leaf])
+
+      do_import_directives(type_module, env, opts)
+    end
+  end
+
+  defp do_import_directives(
+         {{:., _, [{:__aliases__, _, [{:__MODULE__, _, _} | tail]}, :{}]}, _, modules_ast_list},
+         env,
+         opts
+       ) do
+    root_module = Module.concat([env.module | tail])
+
+    for {_, _, leaf} <- modules_ast_list do
+      type_module = Module.concat([root_module | leaf])
+
+      do_import_directives(type_module, env, opts)
+    end
+  end
+
+  defp do_import_directives(
+         {{:., _, [{:__aliases__, _, root}, :{}]}, _, modules_ast_list},
+         env,
+         opts
+       ) do
+    root_module = Module.concat(root)
+    root_module_with_alias = Keyword.get(env.aliases, root_module, root_module)
+
+    for {_, _, leaf} <- modules_ast_list do
+      type_module = Module.concat([root_module_with_alias | leaf])
+
+      do_import_directives(type_module, env, opts)
+    end
+  end
+
+  defp do_import_directives(module, env, opts) do
+    Module.put_attribute(env.module, :__absinthe_directive_imports__, [
+      {module, opts} | Module.get_attribute(env.module, :__absinthe_directive_imports__) || []
+    ])
+
+    []
+  end
+
+  defp do_import_type_extensions(
+         {{:., _, [{:__MODULE__, _, _}, :{}]}, _, modules_ast_list},
+         env,
+         opts
+       ) do
+    for {_, _, leaf} <- modules_ast_list do
+      type_module = Module.concat([env.module | leaf])
+
+      do_import_type_extensions(type_module, env, opts)
+    end
+  end
+
+  defp do_import_type_extensions(
+         {{:., _, [{:__aliases__, _, [{:__MODULE__, _, _} | tail]}, :{}]}, _, modules_ast_list},
+         env,
+         opts
+       ) do
+    root_module = Module.concat([env.module | tail])
+
+    for {_, _, leaf} <- modules_ast_list do
+      type_module = Module.concat([root_module | leaf])
+
+      do_import_type_extensions(type_module, env, opts)
+    end
+  end
+
+  defp do_import_type_extensions(
+         {{:., _, [{:__aliases__, _, root}, :{}]}, _, modules_ast_list},
+         env,
+         opts
+       ) do
+    root_module = Module.concat(root)
+    root_module_with_alias = Keyword.get(env.aliases, root_module, root_module)
+
+    for {_, _, leaf} <- modules_ast_list do
+      type_module = Module.concat([root_module_with_alias | leaf])
+
+      do_import_type_extensions(type_module, env, opts)
+    end
+  end
+
+  defp do_import_type_extensions(module, env, opts) do
+    Module.put_attribute(env.module, :__absinthe_type_extension_imports__, [
+      {module, opts}
+      | Module.get_attribute(env.module, :__absinthe_type_extension_imports__) || []
+    ])
+
+    []
+  end
+
+  @placement {:import_type_extensions, [toplevel: true]}
+  @doc """
+  Import type_extensions from another module
+
+  To selectively import type_extensions you can use the `:only` and `:except` opts.
+
+  ## Placement
+  #{Utils.placement_docs(@placement)}
+
+  ## Examples
+  ```
+  import_type_extensions MyApp.Schema.TypeExtensions
+
+  import_type_extensions MyApp.Schema.TypeExtensions.{TypeExtensionsA, TypeExtensionsB}
+
+  import_type_extensions MyApp.Schema.TypeExtensions, only: [:foo]
+
+  import_type_extensions MyApp.Schema.TypeExtensions, except: [:bar]
+  ```
+  """
+  @doc group: :notation
+  defmacro import_type_extensions(type_module_ast, opts \\ []) do
+    env = __CALLER__
+
+    type_module_ast
+    |> Macro.expand(env)
+    |> do_import_type_extensions(env, opts)
+  end
+
+  @placement {:schema, [toplevel: true, extend: true]}
+  @doc """
+  Declare a schema
+
+  Optional declaration of the schema. Useful if you want to add directives
+  to your schema declaration
+
+  ## Placement
+
+  #{Utils.placement_docs(@placement)}
+
+  ## Examples
+
+  ```
+  schema do
+    directive :feature
+    field :query, :query
+    # ...
+  end
+  ```
+  """
+  @doc group: :notation
+  defmacro schema(do: block) do
+    __CALLER__
+    |> Absinthe.Schema.Notation.recordable!(:schema, @placement[:schema])
+    |> Absinthe.Schema.Notation.record_schema!(block)
   end
 
   defmacro __before_compile__(_) do
